Implements continuous archiving feature.
This commit adds the following commands related to WAL backup and recovery:

1. postgres:wal-enable <name> [-f]

Enables continuous archiving of WAL segments.  This is useful as it allows
incremental backup of the cluster (see
http://www.postgresql.org/docs/9.6/static/continuous-archiving.html).

Archived WAL segments are written to `$SERVICE_ROOT/backups`, which is mounted
in the container as `/var/lib/postgresql/backups`. Users can then backup this
directory on the host using traditional backup mechanisms (rsync, upload to S3,
etc).

If the option '-f' is given, any previous files found at `$SERVICE_ROOT/backups`
will be removed, and a new backup will be created. Otherwise, the command will
fail upon detecting a non-empty `$SERVICE_ROOT/backups` directory.

2. postgres:wal-disable <name>

Disables continuous archiving of WAL segments. This command will NOT remove
previous backup files at `$SERVICE_ROOT/backups`.

3. postgres:wal-recover <name> <timestamp>

Restores the database from a WAL backup created with `postgres:enable-wal`.
Generally, the steps for a disaster recovery is:

a) Create a new database with `postgres:create`.
b) Enable WAL on the new database with `postgres:wal-enable`.
c) Replace the `$SERVICE_ROOT/backups` directory of the newly created database
   with the `backups` directory of the old database.
d) Run `postgres:wal-recover` on the newly created database. After the recovery
   completes, the new database should contain all the data previously written
   to the backup.

If a timestamp (of a format akin to '2016-07-14 22:39:00 EST') is given, the
database will be reverted to/recovered up to the given point in time.
Otherwise, the database will be reverted to/recovered up to the latest point in
time contained in the backup.